SPI, DAC fixes
DAC initialise fn didn't have chip select brought low, so was using a different reference voltage as a result.
SPI was checking equality against 1 when it should have been comparing against non-zero to determine the value of the MOSI line.
Strange issue with smclk, when it's enabled the MSP seems to go into a loop immediately after it's enabled. Switch to using aclk instead (slower but oh well).
Move to relative percent difference instead of accuracy to avoid issues with division by zero.