Prepare for new PCB revision
Make files generic over PCB version using module trickery, now we only need to change the version once in main.rs and it propagates to all other files. 
Also split pcb mapping file into a versioned component and a common component. Maybe a mistake?
Also rename mapping files to make it clear that it's the PCB version that is changing, not the mapping.