Misc changes and fixes
Of particular note, the DAC chip select needs to be set earlier so it can be chosen during it's initialisation. Maybe make this more foolproof in future?
Also needed to screw around with the clock config again. It seems to deterministically (but random with respect to code changes) fail, which results in the MSP clock being stopped(?).