Refactor: Printing responsibilities
For consistency all the test_x_sensor functions return a PerformanceResult, and the x_sensing functions now print them and any other relevant info.
temperature_sensing is an exception since it's a bit more complex (might come back to this), so rename it print_temperatures to make it clear that it prints.